Implement the relay program.
Relay is a small program that connects to a DGD server, logs in as
directed by an input script, and then executes a secondary program
with the connection to the server as stdin and stdout.

Relay is intended for those cases where integrating foreign code into
DGD is hard, because of DGD's deliberate limitations, because of atomic
rollback, or because DGD's internal structure is a bad match for what is
required.